file upload with local browser ocr and indexeddb storage
