+++ client/src/components/App/App.tsx
@@ -1,100 +1,22 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import * as S from "./App.style";
 
 import Layout from "../Layout/Layout";
-import Button from "../Button/Button";
-import { Document } from "../../models";
-import { documentService } from "../../services";
-
-const { createWorker } = require("tesseract.js");
+import FileUpload from "../FileUpload";
 
 function App() {
-  const [selectedFile, setSelectedFile] = useState<{
-    name: string;
-    type: string;
-    size: number;
-    lastModifiedDate: Date;
-  }>();
-  const [document, setDocument] = useState<Document>(new Document());
-  const [isSelected, setIsSelected] = useState(false);
-  const [ocrProgress, setOcrProgress] = useState(0);
-  const [isLoading, setIsLoading] = useState(false);
-  const [documents, setDocuments] = useState([]);
-
-  useEffect(() => {
-    documentService.findDocuments().then((docs) => {
-      console.log(docs);
-    });
-  }, []);
-
-  const changeHandler = (event: any) => {
-    const file = event.target.files[0];
-    setSelectedFile(file);
-
-    let reader = new FileReader();
-    reader.readAsArrayBuffer(file);
-    reader.onload = () => {
-      document.fileBlob = reader.result;
-    };
-
-    document.name = file.name;
-    document.fileType = file.type;
-    document.size = file.size;
-    setDocument(document);
-
-    setIsSelected(true);
-  };
-
-  const handleSubmission = async () => {
-    setIsLoading(true);
-
-    const worker = createWorker({
-      logger: (m: any) => {
-        if (m.status === "recognizing text") {
-          setOcrProgress(m.progress * 100);
-        }
-      },
-    });
-
-    await worker.load();
-    await worker.loadLanguage("eng");
-    await worker.initialize("eng");
-    const {
-      data: { text },
-    } = await worker.recognize(selectedFile);
-    console.log("Recognized text:", text);
-
-    document.extractedText = text;
-    setDocument(document);
-    documentService.addDocument(document);
-
-    setIsSelected(false);
-    setIsLoading(false);
-
-    document.name = await worker.terminate();
-  };
+  // TODO: Example to fuzzy search documents locally
+  // const [documents, setDocuments] = useState([]);
+  // useEffect(() => {
+  //   documentService.findDocuments().then((docs) => {
+  //     console.log("1", docs);
+  //   });
+  // }, []);
 
   return (
     <S.Wrapper>
       <Layout>
-        <div>
-          {isLoading && <p>{`${ocrProgress.toFixed(2)}%`}</p>}
-          <input type="file" name="file" onChange={changeHandler} />
-          {isSelected && selectedFile ? (
-            <div>
-              <p>Filename: {selectedFile?.name}</p>
-              <p>Filetype: {selectedFile?.type}</p>
-              <p>Size in bytes: {selectedFile?.size}</p>
-              <p>
-                lastModifiedDate:{" "}
-                {selectedFile?.lastModifiedDate.toLocaleDateString()}
-              </p>
-              {!isLoading && <Button onClick={handleSubmission}>Submit</Button>}
-            </div>
-          ) : (
-            <p>Select a file to show details</p>
-          )}
-        </div>
+        <FileUpload></FileUpload>
       </Layout>
     </S.Wrapper>
   );
